+++ ingester2/src/persist/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod backpressure;
 pub(super) mod compact;
 mod context;
+pub(crate) mod drain_buffer;
 pub(crate) mod handle;
 pub(crate) mod hot_partitions;
 mod worker;
